product-service/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.23
flask-cors==3.0.10
psycopg2-binary==2.9.1
redis==3.5.3
python-dotenv==0.19.0
pytest==6.2.5 